requirements.txt
requests==2.26.0
python-dotenv==0.19.2
click==8.0.3
